Encapsulate topic finding logic in pipeline.
