+++ pipeline/test_topic.py
@@ -1,0 +1,2 @@
+# pylint: skip-file
+"""Tests for topic module."""
